DDL 에서 auto_increment 속성 삭제, DAO 변경

+++ src/main/java/com/aispace/erksystem/service/database/type/ServiceUserId.java
@@ -9,14 +9,14 @@ import java.util.Objects;
 @Getter
 @Setter
 public class ServiceUserId implements Serializable {
-    private Long userId;
-    private Long orgId;
+    private Integer userId;
+    private Integer orgId;
 
     public ServiceUserId() {
         // 기본 생성자
     }
 
-    public ServiceUserId(long userId, long orgId) {
+    public ServiceUserId(Integer userId, Integer orgId) {
         this.userId = userId;
         this.orgId = orgId;
     }
